renamed the file and removed cooking_screen1
